+++ src/main/java/roomescape/reservation/web/controller/ReservationController.java
@@ -1,4 +1,4 @@
-package roomescape.controller;
+package roomescape.reservation.web.controller;
 
 import java.util.List;
 import lombok.RequiredArgsConstructor;
@@ -10,9 +10,9 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import roomescape.dto.ReservationCreateDto;
-import roomescape.dto.ReservationDto;
-import roomescape.service.ReservationService;
+import roomescape.reservation.service.ReservationService;
+import roomescape.reservation.web.dto.ReservationCreateDto;
+import roomescape.reservation.web.dto.ReservationDto;
 
 @RequiredArgsConstructor
 @RequestMapping("/reservations")
