+++ src/main/java/roomescape/reservationtime/web/dto/ReservationTimeCreateDto.java
@@ -1,7 +1,7 @@
-package roomescape.dto;
+package roomescape.reservationtime.web.dto;
 
 import java.time.LocalTime;
-import roomescape.domain.ReservationTime;
+import roomescape.reservationtime.domain.ReservationTime;
 
 public record ReservationTimeCreateDto(LocalTime startAt) {
